# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PixelSlime API
  version: 1.0.0
  description: |
    Public, read-only API for the PixelSlime SLIMEDEX.

    There is **no authentication**. The site is public and anonymous — per-visitor state
    (which slimes you have opened, theme, reduced motion) lives in `localStorage` and is
    never sent to the server.

    The only write path is the daily Container Apps Job, which is not reachable from the
    internet. `/api/admin/generate` is guarded by a Key Vault admin token header and is off
    by default.

    This document is the **contract** between W6 (frontend) and W7 (backend). W6 generates
    its mock server from this file, so it must never drift.
servers:
  - url: /
...
paths:
...
  /api/nft/{serial}:
    get:
      tags: [meta]
      summary: ERC-721 metadata for the card
      parameters: [{ $ref: "#/components/parameters/Serial" }]
      responses:
        "200":
          description: ERC-721 metadata document
          content:
            application/json:
              schema:
                type: object
                required: [name, description, image, attributes]
                properties:
                  name:        { type: string }
                  description: { type: string }
                  image:       { type: string, format: uri }
                  external_url: { type: string, format: uri }
                  attributes:
                    type: array
                    items:
                      type: object
                      properties:
                        trait_type:   { type: string }
                        value:        {}
                        display_type: { type: string }
        "404": { description: No such card }
  /api/admin/generate:
    post:
      tags: [meta]
      summary: Trigger a generation by hand. Off by default.
      security:
        - adminToken: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                forceRarity: { $ref: "#/components/schemas/Rarity" }
                dryRun:      { type: boolean, default: false }
      responses:
        "202": { description: Generation accepted }
        "401": { description: Missing or wrong admin token }
        "409": { description: A card already exists for today }
components:
  securitySchemes:
    adminToken:
      type: apiKey
      in: header
      name: X-PixelSlime-Admin
  parameters:
    Serial:
      name: serial
      in: path
      required: true
      schema: { type: integer, minimum: 1, maximum: 65535 }
  schemas:
    Rarity:
      type: string
      enum: [COMMON, UNCOMMON, RARE, EPIC, LEGENDARY, MYTHIC]
...